Split project into separate files
All API endpoints are moved to nookipedia/api
