+++ nookipedia/api/__init__.py
@@ -1,0 +1,17 @@
+from nookipedia.api import (
+    admin,
+    art,
+    bugs,
+    clothing,
+    events,
+    fish,
+    furniture,
+    interior,
+    items,
+    photos,
+    recipes,
+    sea,
+    static,
+    tools,
+    villagers
+)
