+++ nookipedia/static/doc.yaml
