+++ nookipedia/static/index.html
