+++ nookipedia/static/js/redoc.standalone.js
